@@ -10,12 +10,13 @@
 //      License: MIT (https://opensource.org/licenses/MIT)
 //
 
+
 namespace System;
 
 /// <summary>Clones the object</summary>
 /// <typeparam name="TSelf">The class' name</typeparam>
 // #if DEFINE_INTERNAL
-internal interface ICloneable<TSelf> where TSelf : ICloneable<TSelf>
+public interface ICloneable<TSelf> where TSelf : ICloneable<TSelf>
 // #else
 // public interface ICloneable<TSelf> where TSelf : ICloneable<TSelf>
 // #endif
