@@ -13,7 +13,7 @@
 namespace System;
 
 // #if DEFINE_INTERNAL
-internal static class StringExtensions
+public static class StringExtensions
 // #else
 // public static class StringExtensions
 // #endif
@@ -25,6 +25,18 @@ internal static class StringExtensions
     {
         if (str is null)
             throw new ArgumentNullException(nameof(str));
-        return str.Replace("&", "\x26").Replace("<", "\x3c").Replace(">", "\x3e").Replace("\"", "\x22").Replace("'", "\x27");
+        return str.Replace("&", @"\x26").Replace("<", @"\x3c").Replace(">", @"\x3e").Replace("\"", @"\x22").Replace("'", @"\x27");
     }
+
+    /// <summary>Determines if a string is null or whitespace</summary>
+    /// <param name="str">The string to see whether it's null or whitespace</param>
+    /// <returns>A <see langword="bool" /> value indicating whether the string was null or whitespace</returns>
+    public static bool IsNullOrWhitespace(this string? str)
+        => string.IsNullOrWhiteSpace(str);
+
+    /// <summary>Determines if a string is null or empty</summary>
+    /// <param name="str">The string to see whether it's null or empty</param>
+    /// <returns>A <see langword="bool" /> value indicating whether the string was null or empty</returns>
+    public static bool IsNullOrEmpty(this string? str)
+        => string.IsNullOrEmpty(str);
 }
